+++ ui/hooks/hardware-wallets/useHwSwapConfirmationMonitoring.test.ts
@@ -1,9 +1,9 @@
 import {
   HardwareWalletSignatureEvent,
   HardwareWalletSignatureStatus,
-} from '../../../pages/hardware-wallets/swap/hardware-wallet-signatures-state-machine';
-import { createSignatureState } from '../../../pages/hardware-wallets/swap/hardware-wallet-signatures-state-machine.test-helpers';
-import { renderHookWithProvider } from '../../../../test/lib/render-helpers-navigate';
+} from '../../pages/hardware-wallets/swap/hardware-wallet-signatures-state-machine';
+import { createSignatureState } from '../../pages/hardware-wallets/swap/hardware-wallet-signatures-state-machine.test-helpers';
+import { renderHookWithProvider } from '../../../test/lib/render-helpers-navigate';
 import { useHwSwapConfirmationMonitoring } from './useHwSwapConfirmationMonitoring';
 
 jest.mock('react-redux', () => ({
