+++ ui/hooks/hardware-wallets/useHwSwapConfirmationMonitoring.ts
@@ -4,9 +4,9 @@ import { useSelector } from 'react-redux';
 import {
   HardwareWalletSignatureEvent,
   HardwareWalletSignatureStatus,
-} from '../../../pages/hardware-wallets/swap/hardware-wallet-signatures-state-machine';
-import type { HardwareWalletSignaturesState } from '../../../pages/hardware-wallets/swap/hardware-wallet-signatures-state-machine';
-import type { BridgeStatusState } from '../../../pages/hardware-wallets/swap/types';
+} from '../../pages/hardware-wallets/swap/hardware-wallet-signatures-state-machine';
+import type { HardwareWalletSignaturesState } from '../../pages/hardware-wallets/swap/hardware-wallet-signatures-state-machine';
+import type { BridgeStatusState } from '../../pages/hardware-wallets/swap/types';
 
 type UseHardwareWalletConfirmationMonitoringOptions = {
   hardwareWalletUsed: boolean;
