+++ ui/hooks/hardware-wallets/useHwSwapConnectionMonitoring.test.ts
@@ -1,14 +1,14 @@
 import { ErrorCode } from '@metamask/hw-wallet-sdk';
-import { ConnectionStatus } from '../../../contexts/hardware-wallets';
+import { ConnectionStatus } from '../../contexts/hardware-wallets';
 import {
   HardwareWalletSignatureEvent,
   HardwareWalletSignatureStatus,
-} from '../../../pages/hardware-wallets/swap/hardware-wallet-signatures-state-machine';
-import { createSignatureState } from '../../../pages/hardware-wallets/swap/hardware-wallet-signatures-state-machine.test-helpers';
-import { renderHookWithProvider } from '../../../../test/lib/render-helpers-navigate';
+} from '../../pages/hardware-wallets/swap/hardware-wallet-signatures-state-machine';
+import { createSignatureState } from '../../pages/hardware-wallets/swap/hardware-wallet-signatures-state-machine.test-helpers';
+import { renderHookWithProvider } from '../../../test/lib/render-helpers-navigate';
 import { useHwSwapConnectionMonitoring } from './useHwSwapConnectionMonitoring';
 
-jest.mock('../../../contexts/hardware-wallets', () => ({
+jest.mock('../../contexts/hardware-wallets', () => ({
   ConnectionStatus: {
     Disconnected: 'disconnected',
     Connecting: 'connecting',
@@ -24,13 +24,13 @@ jest.mock('../../../contexts/hardware-wallets', () => ({
 }));
 
 const mockUseHardwareWalletState = jest.requireMock(
-  '../../../contexts/hardware-wallets',
+  '../../contexts/hardware-wallets',
 ).useHardwareWalletState;
 const mockGetHardwareWalletErrorCode = jest.requireMock(
-  '../../../contexts/hardware-wallets',
+  '../../contexts/hardware-wallets',
 ).getHardwareWalletErrorCode;
 const mockIsUserRejectedHardwareWalletError = jest.requireMock(
-  '../../../contexts/hardware-wallets',
+  '../../contexts/hardware-wallets',
 ).isUserRejectedHardwareWalletError;
 
 describe('useHwSwapConnectionMonitoring', () => {
