+++ ui/hooks/hardware-wallets/useHwSwapConnectionMonitoring.ts
@@ -6,12 +6,12 @@ import {
   getHardwareWalletErrorCode,
   isUserRejectedHardwareWalletError,
   useHardwareWalletState,
-} from '../../../contexts/hardware-wallets';
+} from '../../contexts/hardware-wallets';
 import {
   HardwareWalletSignatureEvent,
   HardwareWalletSignatureStatus,
-} from '../../../pages/hardware-wallets/swap/hardware-wallet-signatures-state-machine';
-import type { HardwareWalletSignaturesState } from '../../../pages/hardware-wallets/swap/hardware-wallet-signatures-state-machine';
+} from '../../pages/hardware-wallets/swap/hardware-wallet-signatures-state-machine';
+import type { HardwareWalletSignaturesState } from '../../pages/hardware-wallets/swap/hardware-wallet-signatures-state-machine';
 
 type UseHardwareWalletConnectionMonitoringOptions = {
   signatureState: HardwareWalletSignaturesState;
