+++ ui/hooks/hardware-wallets/useHwSwapQuoteData.ts
@@ -6,8 +6,8 @@ import {
   getBridgeQuotes,
   getFromToken,
   getToToken,
-} from '../../../ducks/bridge/selectors';
-import { getHardwareWalletType } from '../../../../shared/lib/selectors/keyring';
+} from '../../ducks/bridge/selectors';
+import { getHardwareWalletType } from '../../../shared/lib/selectors/keyring';
 
 /**
  * Provides quote and token data for hardware wallet swap/bridge flows.
