+++ ui/hooks/hardware-wallets/useHwSwapSubmission.test.ts
@@ -6,13 +6,13 @@ import {
 import {
   HardwareWalletSignatureEvent,
   HardwareWalletSignatureStatus,
-} from '../../../pages/hardware-wallets/swap/hardware-wallet-signatures-state-machine';
-import { createSignatureState } from '../../../pages/hardware-wallets/swap/hardware-wallet-signatures-state-machine.test-helpers';
-import { renderHookWithProvider } from '../../../../test/lib/render-helpers-navigate';
-import { rejectPendingApproval } from '../../../store/actions';
+} from '../../pages/hardware-wallets/swap/hardware-wallet-signatures-state-machine';
+import { createSignatureState } from '../../pages/hardware-wallets/swap/hardware-wallet-signatures-state-machine.test-helpers';
+import { renderHookWithProvider } from '../../../test/lib/render-helpers-navigate';
+import { rejectPendingApproval } from '../../store/actions';
 import { useHwSwapSubmission } from './useHwSwapSubmission';
 
-jest.mock('../../../store/actions', () => ({
+jest.mock('../../store/actions', () => ({
   rejectPendingApproval: jest.fn(() => () => Promise.resolve()),
 }));
